debug for warning

+++ DynamicLyrics/LyricX/NCChineseConverter.m
@@ -62,13 +62,13 @@
   //
   NSMutableString *result = [[[NSMutableString alloc] initWithString:@""] autorelease];
   NSDictionary *useDict = _dict[dictTypeKey];
-  int i = 0;
+  NSInteger i = 0;
   while (i < oriString.length) {
       
     NSAutoreleasePool *pool = [[NSAutoreleasePool alloc] init];
 
-    int max = oriString.length - i;
-    int j;
+    NSInteger max = oriString.length - i;
+    NSInteger j;
     for (j = max; j > 0; j--) {
       NSRange range = NSMakeRange(i, j);
       NSString *subStr = [oriString substringWithRange:range];
